--- .eslintignore
@@ -1,3 +1,0 @@
-// .eslintignore
-node_modules
-coverage
