--- .eslintrc
@@ -1,12 +1,0 @@
-// .eslintrc
-{
-	"env": {
-		"browser": true,
-		"es6": true,
-		"jest": true,
-		"node": true
-	},
-	"extends": [
-		"eslint:recommended"
-	]
-}
