+++ src/test.js
@@ -1,0 +1,7 @@
+const getData = async () => {
+  const packageJson = await import("./package.json")
+
+  console.log(packageJson.version)
+}
+
+getData()
